# Remove --fast default from golangci-lint