@jose-elias-alvarez 2023-01-16
I don't know enough about this linter to say for sure, but judging from [the VS Code integration instructions](https://golangci-lint.run/usage/integrations/) it seems like the recommended setup for VS Code does indeed use the `--fast` flag. Are there any other examples?