@BlakeWilliams 2023-01-16
> I don't know enough about this linter to say for sure, but judging from [the VS Code integration instructions](https://golangci-lint.run/usage/integrations/) it seems like the recommended setup for VS Code does indeed use the `--fast` flag. Are there any other examples?

That is true, but it makes me wonder why it's not the default configuration in VSCode. Maybe it's because it's easier to add extra flags than to remove them?

I looked through a few other similar plugins for different editors (on the integration page you linked), as far as I can tell most aren't setting `--fast` as a default, but at least one of them does.