@BlakeWilliams 2023-01-16
For comparisons sake, here's the difference (in line count) I get when using `--fast`.

```
wordup ❯ golangci-lint run | wc -l
     102
wordup ❯ golangci-lint run --fast | wc -l
       3
```