@jose-elias-alvarez 2023-01-17
Hmm, what is the speed difference? We shouldn't have the same problem where the extension blocks the editor, and we only run the linter on save, but it might be good to check if it's going to affect performance.