@academo 2023-01-23
I was just setting up golangci-lint via nulls because I was using the LSP version via lsp-config that gave me problems and the lspconfig verison [doesn't include --fast](https://github.com/neovim/nvim-lspconfig/blob/master/doc/server_configurations.md#golangci_lint_ls) by default either so I was getting different results.

Regarding speed. Here are some tests:

Setup: 
* file of 96KB [>2K lines of code](https://github.com/grafana/grafana/blob/main/pkg/services/ngalert/state/manager_test.
* Use hyperfine for benchmarking

## Tests without cache (notice the prepare)

### without `--fast`
```
❯ hyperfine --runs 10 'GOLANGCI_LINT_CACHE=/tmp/testcache golangci-lint run ./pkg/services/ngalert/state/manager_test.go' --prepare="rm -rf /tmp/testcache;mkdir /tmp/testcache"
Benchmark 1: GOLANGCI_LINT_CACHE=/tmp/testcache golangci-lint run ./pkg/services/ngalert/state/manager_test.go
  Time (mean ± σ):      6.377 s ±  0.040 s    [User: 76.838 s, System: 2.875 s]
  Range (min … max):    6.331 s …  6.477 s    10 runs
```

### with `--fast`
```
❯ hyperfine --runs 10 'GOLANGCI_LINT_CACHE=/tmp/testcache golangci-lint run --fast ./pkg/services/ngalert/state/manager_test.go' --prepare="rm -rf /tmp/testcache;mkdir /tmp/testcache"
Benchmark 1: GOLANGCI_LINT_CACHE=/tmp/testcache golangci-lint run --fast ./pkg/services/ngalert/state/manager_test.go
  Time (mean ± σ):      8.042 s ±  2.054 s    [User: 100.977 s, System: 4.610 s]
  Range (min … max):    6.320 s … 10.790 s    10 runs
```

## Tests with cache (notice the warm up runs):

### Without `--fast`
```
❯ hyperfine --runs 10 'GOLANGCI_LINT_CACHE=/tmp/testcache golangci-lint run ./pkg/services/ngalert/state/manager_test.go' --warmup=2
Benchmark 1: GOLANGCI_LINT_CACHE=/tmp/testcache golangci-lint run ./pkg/services/ngalert/state/manager_test.go
  Time (mean ± σ):     668.1 ms ± 153.9 ms    [User: 1824.5 ms, System: 471.0 ms]
  Range (min … max):   559.4 ms … 949.8 ms    10 runs
```


### With `--fast`
```❯ hyperfine --runs 10 'GOLANGCI_LINT_CACHE=/tmp/testcache golangci-lint run --fast ./pkg/services/ngalert/state/manager_test.go' --warmup=2
Benchmark 1: GOLANGCI_LINT_CACHE=/tmp/testcache golangci-lint run --fast ./pkg/services/ngalert/state/manager_test.go
  Time (mean ± σ):     585.2 ms ±  30.5 ms    [User: 1475.0 ms, System: 451.1 ms]
  Range (min … max):   524.1 ms … 633.2 ms    10 runs
```

## Conclusion

It looks like `--fast` is a bit faster in a cached scenario and worst in a non cached scenario (which correlate to the docs where they mention the first run won't be fast).

Considering this is quite a big file I'd say the difference between `--fast` and non `--fast` is negligible considering the more feedback you get from the non-fast