@jose-elias-alvarez 2023-01-26
Thanks for the detailed breakdown - I agree that the difference isn't significant enough to worry about. Let's merge this and see how it goes.